+++ test/xema/ref_remote_test.exs
@@ -62,9 +62,10 @@ defmodule Xema.RefRemoteTest do
 
   describe "file ref" do
     setup do
-      %{schema: Xema.new({:ref, "integer.exon"})}
+      %{schema: Xema.new({:ref, "integer.exon"}, resolver: Test.FileResolver)}
     end
 
+    @tag :only
     test "validate/2 with a valid value", %{schema: schema} do
       assert validate(schema, 1) == :ok
     end